fix(modules/nixos/fcitx5): fcitx5 started before Wayland compositors
The fcitx5 package provides a XDG autostart file, which is read and
executed by systemd. Thus fcitx5 starts before the Wayland compositor.
If a Wayland environment is not ready, fcitx5 will hang on DBus instead
of exiting. This leads to other applications that depend on DBus,
especially XDG desktop portals, fail to start.

See
<https://discourse.nixos.org/t/fcitx5-failed-to-start-on-wayland/66206>
for relevant information.